alx_travel_app/requirements.txt
Django>=4.2
djangorestframework
drf-yasg
django-cors-headers
celery
django-environ
mysqlclient
